Handle axiom in bootloader mode (#25)
* Formatting and other changes based on Pycharm linter.

* Add project name to file headers.
Tidy up code based on PyCharm linter suggestions.

* Refactor the bootloader get busy calls to be more efficient.
Sending chunks to the bootloader is no longer split between the header and payload. It can all be sent at the same time. The bootloader will automatically handle this.

* Moved the bootloader functions into a class of its own.
Better handling of the device if it is in bootloader mode when the scripts are run.

* Update README.md

* Move convert_device_id_to_string() back into the class.

---------

Co-authored-by: Mark Satterthwaite <mark.satterthwaite@touchnetix.com>