Optionally allow a user to specify a usage read length (#29)
* Formatting and other changes based on Pycharm linter.

* Add project name to file headers.
Tidy up code based on PyCharm linter suggestions.

* Refactor the bootloader get busy calls to be more efficient.
Sending chunks to the bootloader is no longer split between the header and payload. It can all be sent at the same time. The bootloader will automatically handle this.

* Moved the bootloader functions into a class of its own.
Better handling of the device if it is in bootloader mode when the scripts are run.

* Update README.md

* Move convert_device_id_to_string() back into the class.

* Allow read_usage() to take an optional length parameter. This is mainly useful when it comes to reading uF2 whereby the user doesn't necessarily want to read out the full 32KB for data.
Fixed a little bug in u31_DeviceInformation.py where a method call was not correctly updated.

Resolves issue #27.

---------

Co-authored-by: Mark Satterthwaite <mark.satterthwaite@touchnetix.com>